+++ CHANGELOG.md
@@ -8,6 +8,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.4.0] - 2025-08-21
+
+### Added
+
+- `Get-StmClusteredScheduledTaskRun` - Get task runs in a cluster
+- `Stop-StmClusteredScheduledTask` - Stop a running clustered scheduled task
+
 ## [0.3.1] - 2025-08-20
 
 ### Fixed